[Misc] Clean up utils
